backend/requirements.txt
# Platr Backend — Python dependencies
# Install: pip install -r requirements.txt

# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# ── Validation ────────────────────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
google-auth==2.37.0

# ── OSINT / Scraping ─────────────────────────────────────────────────────────
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
selenium==4.27.1
webdriver-manager==4.0.2

# ── LangChain / Multi-Agent ───────────────────────────────────────────────────
langchain==0.3.13
langchain-core==0.3.28
langchain-anthropic==0.3.3
langgraph==0.2.60
anthropic==0.42.0

# ── Email ─────────────────────────────────────────────────────────────────────
aiosmtplib==3.0.1

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
